benchmarks/base/requirements.txt
-r requirements_scenario.txt

pyperf
pyyaml
